feat: add like and dislike functionality to clothing items

+++ controllers/clothingItems.js
@@ -1,40 +1,74 @@
-import ClothingItem from "../models/clothingItem.js";
+import mongoose from "mongoose";
 
-export const getClothingItems = (req, res, next) => {
+import ClothingItem from "../models/clothingItem.js";
+import { InternalServerErrorMessage } from "../utils/errors.js";
+
+export const getClothingItems = (req, res) => {
   ClothingItem.find({})
     .then((clothingItems) => {
       res.send(clothingItems);
     })
     .catch((err) => {
-      next(err);
+      res.status(500).send({ message: InternalServerErrorMessage });
     });
 };
-
-// export const createClothingItem = (req, res, next) => {
-//   ClothingItem.create({
-//     name: req.body.name,
-//     weather: req.body.weather,
-//     imageUrl: req.body.imageUrl,
-//     owner: req.user._id,
-//   })
-//     .then((clothingItem) => {
-//       res.send(clothingItem);
-//     })
-//     .catch((err) => {
-//       next(err);
-//     });
-// };
 
 export const createClothingItem = (req, res) => {
   console.log(req.user._id);
 };
 
-export const deleteClothingItem = (req, res, next) => {
+export const deleteClothingItem = (req, res) => {
+  const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).send({ message: "Invalid user ID." });
+  }
+
   ClothingItem.findByIdAndDelete(req.params.id)
     .then((clothingItem) => {
       res.send(clothingItem);
     })
     .catch((err) => {
-      next(err);
+      res.status(500).send({ message: InternalServerErrorMessage });
+    });
+};
+
+export const likeClothingItem = (req, res) => {
+  const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).send({ message: "Invalid user ID." });
+  }
+
+  ClothingItem.findByIdAndUpdate(
+    req.params.itemId,
+    { $addToSet: { likes: req.user._id } },
+    { new: true }
+  )
+    .then((clothingItem) => {
+      res.send(clothingItem);
+    })
+    .catch((err) => {
+      res.status(500).send({ message: InternalServerErrorMessage });
+    });
+};
+
+export const dislikeClothingItem = (req, res) => {
+  const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).send({ message: "Invalid user ID." });
+  }
+
+  ClothingItem.findByIdAndUpdate(
+    req.params.itemId,
+    { $pull: { likes: req.user._id } },
+    { new: true }
+  )
+    .then((clothingItem) => {
+      res.send(clothingItem);
+    })
+    .catch((err) => {
+      res.status(500).send({ message: InternalServerErrorMessage });
     });
 };
